Refactor themeprovider path
