+++ src/components/Card/CardHeader.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { useMemo } from 'react';
 import PropTypes from 'prop-types';
 
-import { useBootstrapPrefix } from '../ThemeProvider';
+import { useBootstrapPrefix } from '../ThemeProvider/ThemeProvider';
 import CardHeaderContext from './CardHeaderContext';
 import { BsPrefixProps, BsPrefixRefForwardingComponent } from '../helpers';
